Changes to add-gff-info cov_samtools
The behaviour has been changed to decrease the memory requirements, the average coverage is calculated only if `-m` is passed as option and each depth file is scanned one after each other. Moreover, no error is raised if, after finished scanning a depth file, a sequence is not found; instead, a warning with the number of sequences not found is logged. The new assumption is a sequence wasn't found in the depth file, it means that the coverage is 0